fix(portal): split Stepper into separate files for react-refresh
Split the Stepper component files to avoid ESLint react-refresh warnings:
- stepper/types.ts - Type definitions
- stepper/utils.ts - Utility functions (getStepStatus, isStepClickable)
- stepper/useStepper.ts - Hook for state management
- stepper/StepHeader.tsx - Header component
- stepper/Stepper.tsx - Main Stepper component
- stepper/index.ts - Barrel exports
