+++ portal/src/shared/components/index.ts
@@ -6,8 +6,8 @@ export { ProtectedRoute } from './ProtectedRoute'
 export { AdminRoute } from './AdminRoute'
 export { default as DataTable } from './DataTable'
 export { default as ActionMenu } from './ActionMenu'
-export { Stepper, StepHeader, useStepper, getStepStatus, isStepClickable } from './Stepper'
+export { Stepper, StepHeader, useStepper, getStepStatus, isStepClickable } from './stepper'
 export type { BreadcrumbItem } from './Breadcrumbs'
 export type { Column } from './DataTable'
 export type { ActionMenuItem } from './ActionMenu'
-export type { Step, StepStatus, StepperProps, StepHeaderProps } from './Stepper'
+export type { Step, StepStatus, StepperProps, StepHeaderProps } from './stepper'
